+++ src/router/Router.tsx
@@ -6,7 +6,7 @@ import KakaoCallback from '../pages/KakaoCallback/KakaoCallback';
 import Home from '../pages/Home/Home';
 import LayoutWithGNB from '../layouts/LayoutWithGNB';
 import Feed from '../pages/Feed/Feed';
-import Moment from '../pages/Moment/Moment';
+import CreateMoment from '../pages/Moment/CreateMoment';
 import MomentComplete from '../pages/Moment/MomentComplete';
 import MyPage from '../pages/MyPage/MyPage';
 import SelectMode from '../pages/Moment/SelectMode';
@@ -32,8 +32,8 @@ const router = createBrowserRouter([
             element: <SelectMode />,
           },
           {
-            path: '/moment',
-            element: <Moment />,
+            path: '/moment/create-moment',
+            element: <CreateMoment />,
           },
           {
             path: './moment-complete',
